+++ Makefile
@@ -6,6 +6,9 @@ build: node_modules
 watch: node_modules
 	node_modules/.bin/webpack $(WEBPACK_FLAGS) --watch
 
+test: build
+	node_modules/.bin/ava
+
 node_modules: yarn.lock
 	yarn
 	touch $@
